pyproject.toml
[tool.poetry]
name = "tapir"
version = "0.1.0"
description = ""
authors = ["Leon Handreke <leonh@ndreke.de>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "~3.2.13"
django-ldapdb = "^1.5.1"
django-weasyprint = "^1.1.0.post2"
django-extensions = "^3.1.5"
django-bootstrap-datepicker-plus = "^3.0.6"
psycopg2 = "^2.9.3"
django-tables2 = "^2.4.1"
django-filter = "^2.4.0"
tablib = "^3.2.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^5.2.0"}
phonenumbers = "^8.12.44"
weasyprint = "^52.5"
django-select2 = "^7.10.0"
django-silk = "^4.2.0"
gunicorn = "^20.1.0"
django-environ = "^0.8.1"
django-bootstrap5 = "^21.3"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.0.0"
celery = "^5.2.3"
redis = "^4.1.4"
ipython = "^8.1.0"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
ipython = "^8.1.0"
pre-commit = "^2.17.0"
Werkzeug = "^1.0.1"
pytest = "^6.2.5"
pytest-cov = "^2.12.1"
pytest-django = "^4.5.2"
pytest-sugar = "^0.9.4"
selenium = "^3.141.0"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
